authentication issues
spent a good 12 hours trying to get teh authenticator to work. In that tme i managed to do a post request with the email and password which takes the user to the ui. I could not get the admin/user auth to work so will have to keep on trying with implementation