authentication complete
the authentication is now complete. The user sends the details to the web api through the terminal. If verified it gives the user an access token. The user will use that access token to perform curd operations. For CUD, it needs admin verification , so will use hte details from the api call to perforam a select statement and get hte user role (user or admin)